move StatusChangeNotice. extract assets
