+++ core/domain/services/admin/notices/status_change/status_change_notice.css
@@ -1,0 +1,162 @@
+.ee-status-change-notice-wrapper {
+    clear: both;
+    font-size: .8rem;
+    padding-bottom: .75rem;
+}
+
+.ee-status-change-notice__admin-legend {
+    margin: 1rem 0 0;
+}
+
+.ee-status-change-notice__event-editor {
+    margin: 0 0 -2rem;
+    padding: 2rem 0 1rem;
+}
+
+#ee-open-notice-link {
+    box-shadow: none !important;
+    color: var(--ee-status-color-red);
+    display: inline-block;
+    font-weight: 700;
+    margin: .5rem 0;
+    outline: none !important;
+    text-decoration: none;
+}
+
+.ee-close-notice-btn {
+    height: 18px;
+    opacity: 0.4;
+    position: absolute;
+    right: 15px;
+    top: 16px;
+    transition: all 0.15s;
+    width: 18px;
+}
+
+.ee-close-notice-btn:hover {
+    opacity: .8;
+}
+
+.ee-close-notice-btn:after,
+.ee-close-notice-btn:before {
+    background-color: var(--ee-status-color-grey);
+    content: ' ';
+    height: 18px;
+    position: absolute;
+    right: 7px;
+    width: 4px;
+}
+
+.ee-close-notice-btn:after {
+    transform: rotate(-45deg);
+}
+
+.ee-close-notice-btn:before {
+    transform: rotate(45deg);
+}
+
+.ee-hide-container {
+    max-height: 0;
+    overflow: hidden;
+    transition: max-height 0.4s ease;
+}
+
+.ee-open-notice {
+    max-height: 35rem !important;
+}
+
+.ee-show-link {
+    max-height: 1.5rem !important;
+}
+
+.ee-status-change-notice-div {
+    display: inline-block;
+    margin: 0 .25rem 1rem;
+    position: relative;
+    transition: max-height 0.4s ease;
+}
+
+.espresso-events-page .ee-status-message,
+.espresso_events-page .ee-status-message,
+.espresso_messages-page .ee-status-event {
+    display: none;
+}
+
+.ee-status-change-notice-div h3 {
+    color: var(--ee-font-color-light);
+    font-weight: 700;
+    margin: .5rem 0 1rem;
+}
+
+.ee-status-change-notice-div p {
+    font-size: .9rem;
+    margin: .5em 0;
+}
+
+.ee-status-change-notice-div ul {
+    line-height: 2.25rem;
+    list-style: inside;
+}
+
+.ee-status-change-notice-div span.pill {
+    border-radius: 3px;
+    color: white;
+    font-size: .75rem;
+    font-weight: 600;
+    padding: .2rem .6rem;
+}
+
+.ee-status-change-notice-div span.pill.yellow {
+    background: var(--ee-status-color-yellow);
+    color: #403a3a;
+}
+
+.ee-status-change-notice-div span.pill.purple {
+    background: var(--ee-status-color-purple);
+}
+
+.ee-status-change-notice-div span.pill.charcoal {
+    background: var(--ee-status-color-dark-grey);
+}
+
+.ee-status-change-notice-div span.pill.blue {
+    background: var(--ee-status-color-blue);
+}
+
+.ee-status-change-notice-div span.pill.pink {
+    background: var(--ee-status-color-red);
+}
+
+.ee-status-change-notice-div span.pill.green {
+    background: var(--ee-status-color-green);
+}
+
+.ee-status-change-notice-div span.pill.red {
+    background: #fc1922;
+}
+
+.ee-dismiss-notice-pg {
+    text-align: right;
+}
+
+.ee-dismiss-notice-pg a {
+    text-decoration: none;
+}
+
+.ee-status-change-notice-div .ee-dismiss-notice-pg button .pill {
+    background: unset;
+    border: 1px #e65983 solid !important;
+    color: #e65983;
+    transition: all 0.15s;
+}
+
+.ee-status-change-notice-div .ee-dismiss-notice-pg button .pill:hover {
+    background: #e65983;
+    color: #fff;
+}
+
+.ee-status-change-notice-wrapper button {
+    background: none !important;
+    border: none !important;
+    cursor: pointer;
+}
